refactor: change cv bucket file name
